Moved some functions + made clear we won't "weight" CSS x XPath
- Moved some functions to the class where they should fit better (SQLocatorXPath) and
probably make easier to spot some refactoring in the future.
- Also made clear the SQLocator will not "weight" the CSS vs. XPath approaches when
CSS is not enough. Instead of "weighting" (something that was not yet implemented and
just in the wishlist), when the CSS is not able to bring every element the user asked, then
we will user XPath+Filter. The idea of "weighting" was that instead of going straigth to
 XPath+Filter, we'd evaluate if CSS+Filter would not be a better approach.
 From what we know, though, we really don't have means to do such evaluation, and the
 current implementation was much simplified by the XPath+Filter assumption. Well, IF, in
 the future we find some way of doing that weighting, then we refactor it in.